+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
-## 0.0.0 (2023-07-13)
-
+## [0.0.1-beta.0](https://github.com/tushar1998/react-library-boilerplate/compare/v0.0.0-beta.0...v0.0.1-beta.0) (2023-07-23)
 
 ### Features
 
-* **library:** initial boilerplate ([9ff2187](https://github.com/tushar1998/preact-library-boilerplate/commit/9ff21875149c19b4695bdf0876e56e362f727393))
+- **components:** new component button ([cff354b](https://github.com/tushar1998/react-library-boilerplate/commit/cff354b22ffe2f0871211c7ce30086d53e4c551d))
+
+## [0.0.0-beta.0](https://github.com/tushar1998/react-library-boilerplate/compare/v0.0.0...v0.0.0-beta.0) (2023-07-23)
+
+## 0.0.0 (2023-07-13)
+
+### Features
+
+- **library:** initial boilerplate ([9ff2187](https://github.com/tushar1998/preact-library-boilerplate/commit/9ff21875149c19b4695bdf0876e56e362f727393))